Add files via upload
Second sprint of the UNCC faculty connections webscraper. I did work on the engineering department and the language and physics departments. This is only the work I did as my group spread out the webscraping of every UNCC department among us. The final product containing all classes and full functionality, including the ability to be automatically imported to the faculty connections website, is owned by the school and I do not have direct access to it.